src/requirements.txt
Flask==1.1.2
flask-restplus==0.13.0
Flask-SQLAlchemy==2.5.1
Jinja2==3.0.1
pytest==6.2.4
pytest-flask==1.2.0
six==1.16.0
SQLAlchemy==1.4.17
Werkzeug==0.16.0
PyYAML==5.4.1
